refactor(registration-react): Move App component to root directory
