+++ applications/registration-react/src/app/connected-app.jsx
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import _ from 'lodash';
 
 import { App } from './app';
-import { loadConfigFromServer } from '../../redux/modules/config';
+import { loadConfigFromServer } from '../redux/modules/config';
 
 const mapStateToProps = ({ config }) => ({
   registrationLanguage: _.get(config, 'registrationLanguage', 'nb')
