+++ applications/registration-react/src/index.jsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 
 import { configureStore } from './redux/configureStore';
 import { ConnectedFeatureToggleProvider } from './components/feature-toggle/connected-feature-toggle-provider';
-import { ConnectedApp } from './components/app/connected-app';
+import { ConnectedApp } from './app/connected-app';
 
 const store = configureStore();
